Разбей модуль на 2, для того, чтобы не тащить за собой ненужный стек
